mom_track3/code/requirements.txt
beautifulsoup4==4.10.0
folium==0.12.1
GDAL==3.1.4
geopandas==0.8.2
matplotlib==3.3.2
numpy==1.19.2
opencv_python==4.4.0.46
osgeo==0.0.1
pandas==1.1.3
Pillow==8.4.0
pyproj==3.0.0.post1
rasterio==1.2.0
requests==2.25.1
sentinelsat==0.14
Shapely==1.7.1
